+++ CHANGELOG.md
@@ -4,13 +4,21 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `orange2` and `orange3` to colors config.
-- Fix: `VerticalProgress`: Fix `z-index` value.
-- Feature: Add `neon` modifier to Button component.
+## [3.14.0] - 2021-05-25
+
+Features:
+
+- Add `orange2` and `orange3` to colors config.
+- `Button`: Add `neon` modifier.
+
+Fix:
+
+- `VerticalProgress`: Fix `z-index` value.
 
 ## [3.13.1] - 2021-05-20
  
 Fixes:
+
 - `VerticalProgress` : Export component.
 - `DashboardLayout`: Specify size for `DashboardLayout.Flow` content column.